+++ src/main/java/faang/school/achievement/handler/AbstractAchievementHandler.java
@@ -5,7 +5,6 @@ import faang.school.achievement.model.AchievementProgress;
 import faang.school.achievement.event.AlbumCreatedEvent;
 import faang.school.achievement.model.UserAchievement;
 import faang.school.achievement.service.AchievementCache;
-import faang.school.achievement.service.AchievementProgressService;
 import faang.school.achievement.service.AchievementService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -19,16 +18,15 @@ public abstract class AbstractAchievementHandler implements EventHandler<AlbumCr
 
     protected final AchievementCache achievementCache;
     protected final AchievementService achievementService;
-    protected final AchievementProgressService achievementProgressService;
 
     @Async("eventHandlingTaskExecutor")
     public void handleAchievement(AlbumCreatedEvent event, String titleAchievement) {
         Achievement achievement = achievementCache.get(titleAchievement);
         if (!achievementService.hasAchievement(event.getUserId(), achievement.getId())) {
-            achievementProgressService.createProgressIfNecessary(event.getUserId(), achievement.getId());
-            AchievementProgress achievementProgress = achievementProgressService.getProgress(event.getUserId(), achievement.getId());
+            achievementService.createProgressIfNecessary(event.getUserId(), achievement.getId());
+            AchievementProgress achievementProgress = achievementService.getProgress(event.getUserId(), achievement.getId());
             achievementProgress.increment();
-            achievementProgressService.saveProgress(achievementProgress);
+            achievementService.saveProgress(achievementProgress);
             if (achievementProgress.getCurrentPoints() >= achievement.getPoints()) {
                 UserAchievement userAchievement = new UserAchievement();
                 userAchievement.setUserId(event.getUserId());
