+++ src/main/java/faang/school/achievement/handler/LibrarianAchievementHandler.java
@@ -2,7 +2,6 @@ package faang.school.achievement.handler;
 
 import faang.school.achievement.event.AlbumCreatedEvent;
 import faang.school.achievement.service.AchievementCache;
-import faang.school.achievement.service.AchievementProgressService;
 import faang.school.achievement.service.AchievementService;
 import org.springframework.stereotype.Service;
 
@@ -10,9 +9,8 @@ import org.springframework.stereotype.Service;
 public class LibrarianAchievementHandler extends AbstractAchievementHandler {
 
     public LibrarianAchievementHandler(AchievementCache achievementCache,
-                                       AchievementService achievementService,
-                                       AchievementProgressService achievementProgressService) {
-        super(achievementCache, achievementService, achievementProgressService);
+                                       AchievementService achievementService) {
+        super(achievementCache, achievementService);
     }
 
     @Override
