+++ src/test/java/faang/school/achievement/service/AchievementServiceTest.java
@@ -1,9 +1,12 @@
 package faang.school.achievement.service;
 
 import faang.school.achievement.dto.AchievementDto;
+import faang.school.achievement.exception.EntityNotFoundException;
 import faang.school.achievement.mapper.achievement.AchievementMapper;
 import faang.school.achievement.model.Achievement;
+import faang.school.achievement.model.AchievementProgress;
 import faang.school.achievement.model.UserAchievement;
+import faang.school.achievement.repository.AchievementProgressRepository;
 import faang.school.achievement.repository.UserAchievementRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -14,8 +17,10 @@ import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.List;
+import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -26,6 +31,9 @@ class AchievementServiceTest {
 
     @Mock
     private UserAchievementRepository userAchievementRepository;
+
+    @Mock
+    private AchievementProgressRepository achievementProgressRepository;
 
     @Spy
     private AchievementMapper achievementMapper = Mappers.getMapper(AchievementMapper.class);
@@ -83,5 +91,47 @@ class AchievementServiceTest {
         achievementService.giveAchievement(userAchievement);
 
         verify(userAchievementRepository, times(1)).save(userAchievement);
+    }
+
+    @Test
+    void createProgressIfNecessaryTest() {
+        long userId = 1L;
+        long achievementId = 2L;
+
+        achievementService.createProgressIfNecessary(userId, achievementId);
+
+        verify(achievementProgressRepository, times(1)).createProgressIfNecessary(userId, achievementId);
+    }
+
+    @Test
+    void getProgressThrowTest() {
+        long userId = 1L;
+        long achievementId = 2L;
+
+        assertThrows(EntityNotFoundException.class, () -> achievementService.getProgress(userId, achievementId));
+    }
+
+    @Test
+    void getProgressTest() {
+        long userId = 1L;
+        long achievementId = 2L;
+        AchievementProgress achievementProgress = new AchievementProgress();
+        when(achievementProgressRepository.findByUserIdAndAchievementId(userId, achievementId))
+                .thenReturn(Optional.of(achievementProgress));
+
+        achievementService.getProgress(userId, achievementId);
+
+        verify(achievementProgressRepository, times(1)).findByUserIdAndAchievementId(userId, achievementId);
+    }
+
+    @Test
+    void saveProgressTest() {
+        AchievementProgress achievementProgress = new AchievementProgress();
+        achievementProgress.setUserId(1L);
+        achievementProgress.setAchievement(new Achievement());
+
+        achievementService.saveProgress(achievementProgress);
+
+        verify(achievementProgressRepository, times(1)).save(achievementProgress);
     }
 }
